Move featured beers to own route; create placeholder for home page
