+++ client/src/routes/Home.tsx
@@ -1,52 +1,12 @@
-import { useQuery } from "@apollo/client";
-import { Container, Grid, Title, Loader, Alert, Center } from "@mantine/core";
-import { FEATURED_BEERS } from "../graphql/queries";
-import { Beer } from "../types";
-import BeerCard from "../components/BeerCard";
-import { usePageTitle } from "../hooks/usePageTitle";
-
-type FeaturedBeersResult = {
-  featuredBeers: Beer[];
-};
+import { Container, Title } from "@mantine/core";
 
 export default function Home() {
-  usePageTitle({ title: "Discover Great Beer" });
-
-  const { loading, error, data } =
-    useQuery<FeaturedBeersResult>(FEATURED_BEERS);
-
-  if (loading) {
-    return (
-      <Container mt="xl">
-        <Center>
-          <Loader />
-        </Center>
-      </Container>
-    );
-  }
-
-  if (error) {
-    return (
-      <Container mt="xl">
-        <Alert color="red" title="Error">
-          Error loading beers: {error.message}
-        </Alert>
-      </Container>
-    );
-  }
-
   return (
     <Container mt="xl">
       <Title order={1} mb="lg">
-        Featured Beers
+        Home!
       </Title>
-      <Grid gutter="lg">
-        {data?.featuredBeers?.map((beer) => (
-          <Grid.Col key={beer.id} span={{ base: 12, sm: 6, md: 4 }}>
-            <BeerCard beer={beer} />
-          </Grid.Col>
-        ))}
-      </Grid>
+      Coming Soon!
     </Container>
   );
 }
